perf: Don't build in the payload directory on helix
A copy of the `performance` repo is sent as payload to helix. And then
we build from the same directory. But this is shared by the whole job.
And it gets uploaded only once. So, all the workitems on that machine
would share that directory.

This can cause issues like if a workitem has to be retried by helix,
then it will end up building incrementally, because it is using the same
artifacts from the first build. This case was hit when some wasm/aot
workitems were retries.

Solution: copy the diretory to `$HELIX_WORKITEM_ROOT/performance` before
building, since `$HELIX_WORKITEM_ROOT` is private to a workitem.

Thanks to @premun for the suggestion.